refactor: background-handlers.jsをclean-architecture構造に移動
restricted-imports/background-handlers.js を
clean-architecture/infrastructure/browser/handlers/background.js に移動。
クリーンアーキテクチャの層構造に対応した配置に変更。
