+++ host-frontend-root/frontend-src-root/eslint-rules/clean-architecture/infrastructure/browser/handlers/background.js
