+++ host-frontend-root/frontend-src-root/eslint.config.js
@@ -9,7 +9,7 @@ import typescript from './eslint-rules/typescript.js';
 import react from './eslint-rules/react.js';
 import general from './eslint-rules/general.js';
 import noRelativePaths from './eslint-rules/no-relative-paths.js';
-import backgroundHandlers from './eslint-rules/restricted-imports/background-handlers.js';
+import backgroundHandlers from './eslint-rules/clean-architecture/infrastructure/browser/handlers/background.js';
 import contentHandlers from './eslint-rules/restricted-imports/content-handlers.js';
 import contentUsecases from './eslint-rules/restricted-imports/content-usecases.js';
 import domainLayer from './eslint-rules/clean-architecture/domain.js';
